Changed autodiff to dynamicautodiff

+++ include/init-helpers.h
@@ -15,9 +15,9 @@
 // TODO KG Normalised part
 struct total_data {
   cv::Mat base_img;
-  Eigen::MatrixXf base_img_normalised;
+  Eigen::MatrixXd base_img_normalised;
   std::vector<cv::Mat> frames;
-  std::vector<Eigen::MatrixXf> normalised_frames;
+  std::vector<Eigen::MatrixXd> normalised_frames;
 
   total_data() {};
   total_data(std::string inp_dir, int num_images) {
